+++ packages/extension-base/src/services/transaction-service/utils.ts
@@ -34,6 +34,10 @@ function getBlockExplorerAccountRoute (explorerLink: string) {
     return 'accounts';
   }
 
+  if (explorerLink.includes('mainnet-gw2.mosaicchain.io')) {
+    return 'accounts';
+  }
+
   if (explorerLink.includes('statescan.io')) {
     return '#/accounts';
   }
@@ -71,7 +75,7 @@ function getBlockExplorerAccountRoute (explorerLink: string) {
   }
 
   if (explorerLink.includes('node.xode.net')) {
-    return 'account';
+    return 'polkadot/account';
   }
 
   if (explorerLink.includes('tonviewer.com')) {
@@ -110,7 +114,11 @@ function getBlockExplorerTxRoute (chainInfo: _ChainInfo) {
     return 'extrinsics';
   }
 
-  if (['mosaicTest', 'polkadex'].includes(chainInfo.slug)) {
+  if (['xode'].includes(chainInfo.slug)) {
+    return 'polkadot/extrinsics';
+  }
+
+  if (['mosaicTest', 'polkadex', 'mosaic'].includes(chainInfo.slug)) {
     return 'transactions';
   }
 
@@ -162,10 +170,6 @@ export function getExplorerLink (chainInfo: _ChainInfo, value: string, type: 'ac
 
     if (chainInfo.slug === 'tangle') {
       return (`${explorerLink}${explorerLink.endsWith('/') ? '' : '/'}extrinsic/${value}${route}/${value}`);
-    }
-
-    if (chainInfo.slug === 'xode') {
-      return (`${explorerLink}${explorerLink.endsWith('/') ? '' : '/'}polkadot-chain-transaction?search=${value}`);
     }
 
     if (['truth_network', 'aventus'].includes(chainInfo.slug)) {