+++ packages/extension-koni-ui/src/Popup/Confirmations/variants/Transaction/variants/TransferBlock.tsx
@@ -3,11 +3,16 @@
 
 import { ExtrinsicDataTypeMap, ExtrinsicType } from '@subwallet/extension-base/background/KoniTypes';
 import { _isAcrossChainBridge } from '@subwallet/extension-base/services/balance-service/transfer/xcm/acrossBridge';
-import { AlertBox } from '@subwallet/extension-koni-ui/components';
+import { isSubstrateCrossChain } from '@subwallet/extension-base/services/balance-service/transfer/xcm/utils';
+import { _getAssetDecimals, _getAssetPriceId, _getAssetSymbol, _getChainName } from '@subwallet/extension-base/services/chain-service/utils';
+import { AlertBox, CommonTransactionInfo } from '@subwallet/extension-koni-ui/components';
 import MetaInfo from '@subwallet/extension-koni-ui/components/MetaInfo/MetaInfo';
 import QuoteRateDisplay from '@subwallet/extension-koni-ui/components/Swap/QuoteRateDisplay';
-import { useGetNativeTokenBasicInfo } from '@subwallet/extension-koni-ui/hooks';
+import { BN_TEN } from '@subwallet/extension-koni-ui/constants';
+import { useGetAccountByAddress, useGetNativeTokenBasicInfo } from '@subwallet/extension-koni-ui/hooks';
 import { RootState } from '@subwallet/extension-koni-ui/stores';
+import { Number } from '@subwallet/react-ui';
+import BigN from 'bignumber.js';
 import CN from 'classnames';
 import React, { useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
@@ -22,6 +27,8 @@ const Component: React.FC<Props> = ({ className, transaction }: Props) => {
   const { t } = useTranslation();
   const data = transaction.data as ExtrinsicDataTypeMap[ExtrinsicType.TRANSFER_BALANCE];
   const xcmData = transaction.data as ExtrinsicDataTypeMap[ExtrinsicType.TRANSFER_XCM];
+  const priceMap = useSelector((root: RootState) => root.price.priceMap);
+  const { currencyData } = useSelector((root: RootState) => root.price);
   const chainInfoMap = useSelector((root: RootState) => root.chainStore.chainInfoMap);
   const assetRegistryMap = useSelector((root: RootState) => root.assetRegistry.assetRegistry);
   const tokenInfo = assetRegistryMap[transaction.extrinsicType === ExtrinsicType.TRANSFER_XCM ? xcmData.tokenSlug : data.tokenSlug];
@@ -38,59 +45,150 @@ const Component: React.FC<Props> = ({ className, transaction }: Props) => {
     return tokenInfo;
   }, [isAcrossBridge, xcmData.metadata?.destChainSlug, tokenInfo, assetRegistryMap]);
 
-  const chainInfo = useMemo(
-    () => chainInfoMap[transaction.chain],
-    [chainInfoMap, transaction.chain]
-  );
-
+  const fromAccount = useGetAccountByAddress(data.from);
+  const toAccount = useGetAccountByAddress(data.to);
+  const fromAccountName = useMemo(() => fromAccount?.name, [fromAccount]);
+  const toAccountName = useMemo(() => toAccount?.name, [toAccount]);
+  const destinationChainSlug = useMemo(() => xcmData?.destinationNetworkKey || transaction.chain, [xcmData?.destinationNetworkKey, transaction.chain]);
+  const originChainSlug = useMemo(() => xcmData?.originNetworkKey || transaction.chain, [xcmData?.originNetworkKey, transaction.chain]);
+  const senderLabel = useMemo(() => fromAccount?.isMultisig ? t('ui.TRANSACTION.Confirmations.TransferBlock.multisig') : t('ui.TRANSACTION.Confirmations.TransferBlock.sender'), [fromAccount?.isMultisig, t]);
+  const chainInfo = useMemo(() => chainInfoMap[transaction.chain], [chainInfoMap, transaction.chain]);
   const { decimals: nativeTokenDecimals, symbol: nativeTokenSymbol } = useGetNativeTokenBasicInfo(transaction.chain);
   const feeInfo = transaction.estimateFee;
+  const crossChainFeeInfo = transaction?.xcmDestinationFee;
+
+  const priceNativeValue = useMemo(() => {
+    const nativeTokenSlug = `${transaction.chain}-NATIVE-${nativeTokenSymbol}`;
+    const nativeTokenInfo = assetRegistryMap[nativeTokenSlug];
+
+    const priceNativeId = _getAssetPriceId(nativeTokenInfo);
+
+    return priceMap[priceNativeId] || 0;
+  }, [assetRegistryMap, nativeTokenSymbol, priceMap, transaction.chain]);
+
+  const transferTokenSymbol = _getAssetSymbol(tokenInfo);
+  const transferTokenDecimals = _getAssetDecimals(tokenInfo);
+
+  const transferTokenValue = useMemo(() => {
+    const transferTokenPriceId = _getAssetPriceId(tokenInfo);
+
+    return priceMap[transferTokenPriceId] || 0;
+  }, [priceMap, tokenInfo]);
+
+  const destTransferTokenValue = useMemo(() => {
+    const destTransferTokenPriceId = _getAssetPriceId(destTokenInfo);
+
+    return priceMap[destTransferTokenPriceId] || 0;
+  }, [destTokenInfo, priceMap]);
+
+  const convertedFeeValueToUSD = useMemo(() => {
+    if (!feeInfo?.value) {
+      return 0;
+    }
+
+    return new BigN(feeInfo?.value)
+      .multipliedBy(priceNativeValue)
+      .dividedBy(BN_TEN.pow(nativeTokenDecimals || 0))
+      .toNumber();
+  }, [feeInfo, nativeTokenDecimals, priceNativeValue]);
+
+  const convertedCrossChainFeeValueToUSD = useMemo(() => {
+    if (!crossChainFeeInfo?.value) {
+      return 0;
+    }
+
+    return new BigN(crossChainFeeInfo?.value)
+      .multipliedBy(transferTokenValue)
+      .dividedBy(BN_TEN.pow(tokenInfo.decimals || 0))
+      .toNumber();
+  }, [crossChainFeeInfo, tokenInfo, transferTokenValue]);
+
+  const convertedXcmDestTokenValueToUSD = useMemo(() => {
+    if (!isAcrossBridge || !xcmData.metadata?.amountOut) {
+      return 0;
+    }
+
+    return new BigN(xcmData.metadata?.amountOut)
+      .multipliedBy(destTransferTokenValue)
+      .dividedBy(BN_TEN.pow(destTokenInfo.decimals || 0))
+      .toNumber();
+  }, [destTokenInfo.decimals, destTransferTokenValue, isAcrossBridge, xcmData.metadata?.amountOut]);
+
+  const isSubstrateXcm = useMemo(() => {
+    const originChainInfo = chainInfoMap[originChainSlug];
+    const destinationChainInfo = chainInfoMap[destinationChainSlug];
+
+    if (!originChainInfo || !destinationChainInfo) {
+      return false;
+    }
+
+    return isSubstrateCrossChain(originChainInfo, destinationChainInfo);
+  }, [destinationChainSlug, originChainSlug, chainInfoMap]);
 
   return (
     <>
-      <MetaInfo hasBackgroundWrapper>
-        <MetaInfo.Account
-          address={data.from}
-          label={t('ui.TRANSACTION.Confirmations.TransferBlock.sendFrom')}
-        />
-
-        {
-          transaction.extrinsicType === ExtrinsicType.TRANSFER_XCM && chainInfo &&
-          (
-            <MetaInfo.Chain
-              chain={chainInfo.slug}
-              label={t('ui.TRANSACTION.Confirmations.TransferBlock.senderNetwork')}
+      {
+        transaction.wrappingStatus
+          ? (
+            <CommonTransactionInfo
+              address={data.from}
+              network={transaction.chain}
             />
           )
+          : (!!chainInfo && transaction.extrinsicType === ExtrinsicType.TRANSFER_XCM
+            ? (
+              <MetaInfo hasBackgroundWrapper>
+                <MetaInfo.Transfer
+                  alwaysShowChain
+                  destinationChain={{
+                    slug: destinationChainSlug,
+                    name: _getChainName(chainInfoMap[destinationChainSlug])
+                  }}
+                  originChain={{
+                    slug: originChainSlug,
+                    name: _getChainName(chainInfoMap[originChainSlug])
+                  }}
+                  recipientAddress={data.to}
+                  recipientLabel={t('ui.TRANSACTION.Confirmations.TransferBlock.recipient')}
+                  recipientName={toAccountName}
+                  senderAddress={data.from}
+                  senderLabel={senderLabel}
+                  senderName={fromAccountName}
+                />
+              </MetaInfo>
+            )
+            : (
+              <MetaInfo hasBackgroundWrapper>
+                <MetaInfo.Account
+                  address={data.from}
+                  label={t('ui.TRANSACTION.Confirmations.TransferBlock.sendFrom')}
+                />
+                <MetaInfo.Account
+                  address={data.to}
+                  label={t('ui.TRANSACTION.Confirmations.TransferBlock.sendTo')}
+                />
+                <MetaInfo.Chain
+                  chain={chainInfo.slug}
+                  label={t('ui.TRANSACTION.Confirmations.TransferBlock.network')}
+                />
+              </MetaInfo>
+            ))
+      }
+
+      <MetaInfo
+        className={className}
+        hasBackgroundWrapper
+        labelColorScheme={'gray'}
+        valueColorScheme={'light'}
+      >
+        {!!transaction.wrappingStatus &&
+          <MetaInfo.Account
+            address={data.to}
+            chainSlug={transaction.chain}
+            label={t('ui.TRANSACTION.Confirmations.TransferBlock.recipient')}
+            onlyShowName
+          />
         }
-
-        <MetaInfo.Account
-          address={data.to}
-          label={t('ui.TRANSACTION.Confirmations.TransferBlock.sendTo')}
-        />
-
-        {
-          transaction.extrinsicType === ExtrinsicType.TRANSFER_XCM && chainInfo &&
-          (
-            <MetaInfo.Chain
-              chain={xcmData.destinationNetworkKey}
-              label={t('ui.TRANSACTION.Confirmations.TransferBlock.destinationNetwork')}
-            />
-          )
-        }
-
-        {
-          transaction.extrinsicType !== ExtrinsicType.TRANSFER_XCM && chainInfo &&
-          (
-            <MetaInfo.Chain
-              chain={chainInfo.slug}
-              label={t('ui.TRANSACTION.Confirmations.TransferBlock.network')}
-            />
-          )
-        }
-      </MetaInfo>
-
-      <MetaInfo hasBackgroundWrapper>
         {isAcrossBridge && xcmData.metadata
           ? <>
             <MetaInfo.Default
@@ -99,32 +197,103 @@ const Component: React.FC<Props> = ({ className, transaction }: Props) => {
               <QuoteRateDisplay
                 className={'__quote-estimate-swap-value'}
                 fromAssetInfo={tokenInfo}
-                rateValue={Number(xcmData.metadata.rate)}
+                rateValue={parseFloat(xcmData.metadata.rate)}
                 toAssetInfo={destTokenInfo}
               />
             </MetaInfo.Default>
-            <MetaInfo.Number
-              decimals={destTokenInfo.decimals || 0}
-              label={t('ui.TRANSACTION.Confirmations.TransferBlock.expectedAmount')}
-              suffix={destTokenInfo.symbol}
-              value={xcmData.metadata.amountOut}
-            />
+            <div className={'__row -type-default __fee-row'}>
+              <div className={'__field-line-1'}>
+                <div className={'__label'}>
+                  {t('ui.TRANSACTION.Confirmations.TransferBlock.expectedAmount')}
+                </div>
+
+                <div className={'__value'}>
+                  <Number
+                    decimal={destTokenInfo.decimals || 0}
+                    suffix={destTokenInfo.symbol}
+                    value={xcmData.metadata.amountOut}
+                  />
+                </div>
+              </div>
+
+              <div className={'__field-line-2'}>
+                <Number
+                  className={'__fee-price-value'}
+                  decimal={0}
+                  prefix={`~ ${(currencyData.isPrefix && currencyData.symbol) || ''}`}
+                  suffix={(!currencyData.isPrefix && currencyData.symbol) || ''}
+                  value={convertedXcmDestTokenValueToUSD}
+                />
+              </div>
+            </div>
           </>
           : (
-            <MetaInfo.Number
-              decimals={tokenInfo.decimals || 0}
+            <MetaInfo.Default
               label={t('ui.TRANSACTION.Confirmations.TransferBlock.amount')}
-              suffix={tokenInfo.symbol}
-              value={data.value || 0}
-            />
+            >
+              <Number
+                decimal={tokenInfo.decimals || 0}
+                suffix={tokenInfo.symbol}
+                value={data.value || 0}
+              />
+            </MetaInfo.Default>
           )}
 
-        <MetaInfo.Number
-          decimals={feeInfo ? feeInfo.decimals : nativeTokenDecimals}
-          label={t('ui.TRANSACTION.Confirmations.TransferBlock.estimatedFee')}
-          suffix={feeInfo ? feeInfo.symbol : nativeTokenSymbol}
-          value={feeInfo ? feeInfo.value : 0}
-        />
+        {!transaction.wrappingStatus && (
+          <div className={'__row -type-default __fee-row'}>
+            <div className={'__field-line-1'}>
+              <div className={'__label'}>
+                {t('ui.TRANSACTION.Confirmations.TransferBlock.networkFee')}
+              </div>
+
+              <div className={'__value'}>
+                <Number
+                  decimal={feeInfo ? feeInfo.decimals : nativeTokenDecimals}
+                  suffix={feeInfo ? feeInfo.symbol : nativeTokenSymbol}
+                  value={feeInfo ? feeInfo.value : 0}
+                />
+              </div>
+            </div>
+
+            <div className={'__field-line-2'}>
+              <Number
+                className={'__fee-price-value'}
+                decimal={0}
+                prefix={`~ ${(currencyData.isPrefix && currencyData.symbol) || ''}`}
+                suffix={(!currencyData.isPrefix && currencyData.symbol) || ''}
+                value={convertedFeeValueToUSD}
+              />
+            </div>
+          </div>
+        )}
+
+        {transaction.extrinsicType === ExtrinsicType.TRANSFER_XCM && crossChainFeeInfo?.value && new BigN(crossChainFeeInfo.value).gt(0) && isSubstrateXcm && (
+          <div className={'__row -type-default __fee-row'}>
+            <div className={'__field-line-1'}>
+              <div className={'__label'}>
+                {t('ui.TRANSACTION.Confirmations.TransferBlock.crossChainFee')}
+              </div>
+
+              <div className={'__value'}>
+                <Number
+                  decimal={transferTokenDecimals}
+                  suffix={transferTokenSymbol}
+                  value={crossChainFeeInfo ? crossChainFeeInfo?.value : 0}
+                />
+              </div>
+            </div>
+
+            <div className={'__field-line-2'}>
+              <Number
+                className={'__fee-price-value'}
+                decimal={0}
+                prefix={`~ ${(currencyData.isPrefix && currencyData.symbol) || ''}`}
+                suffix={(!currencyData.isPrefix && currencyData.symbol) || ''}
+                value={convertedCrossChainFeeValueToUSD}
+              />
+            </div>
+          </div>
+        )}
       </MetaInfo>
       {
         transaction.extrinsicType === ExtrinsicType.TRANSFER_XCM &&
@@ -143,6 +312,27 @@ const Component: React.FC<Props> = ({ className, transaction }: Props) => {
 
 export const TransferBlock = styled(Component)<Props>(({ theme: { token } }: Props) => {
   return {
+    '.__fee-price-value': {
+      fontSize: token.fontSizeSM,
+      lineHeight: '20px',
+      color: token.colorTextTertiary
+    },
+    '.__fee-row': {
+      flexDirection: 'column',
+      overflow: 'visible'
+    },
+    '.__field-line-1': {
+      display: 'flex',
+      alignItems: 'center',
+      justifyContent: 'space-between',
+      gap: token.sizeXXS,
+      width: '100%'
+    },
+    '.__field-line-2': {
+      display: 'flex',
+      justifyContent: 'flex-end',
+      width: '100%'
+    },
     '&.alert-area': {
       marginTop: token.marginSM
     }