+++ packages/extension-koni-ui/src/utils/transaction/detectType.ts
@@ -16,7 +16,8 @@ export const isTypeStaking = (txType: ExtrinsicType) => [
   ExtrinsicType.STAKING_UNBOND,
   ExtrinsicType.STAKING_WITHDRAW,
   ExtrinsicType.STAKING_COMPOUNDING,
-  ExtrinsicType.STAKING_CANCEL_UNSTAKE
+  ExtrinsicType.STAKING_CANCEL_UNSTAKE,
+  ExtrinsicType.CHANGE_BITTENSOR_ROOT_CLAIM_TYPE
 ].includes(txType);
 
 export const isTypeMint = (txType: ExtrinsicType) =>
@@ -49,4 +50,16 @@ export const isTypeGov = (txType: ExtrinsicType) => [
   ExtrinsicType.GOV_VOTE,
   ExtrinsicType.GOV_UNVOTE,
   ExtrinsicType.GOV_UNLOCK_VOTE
+].includes(txType);
+
+export const isTypeManageSubstrateProxy = (txType: ExtrinsicType) => [
+  ExtrinsicType.ADD_SUBSTRATE_PROXY_ACCOUNT,
+  ExtrinsicType.REMOVE_SUBSTRATE_PROXY_ACCOUNT
+].includes(txType);
+
+export const isTypeMultisig = (txType: ExtrinsicType) => [
+  ExtrinsicType.MULTISIG_APPROVE_TX,
+  ExtrinsicType.MULTISIG_EXECUTE_TX,
+  ExtrinsicType.MULTISIG_CANCEL_TX,
+  ExtrinsicType.MULTISIG_INIT_TX
 ].includes(txType);