+++ CHANGELOG.md
@@ -1,5 +1,90 @@
 # CHANGELOG
 
+## 1.3.78
+Build date: May 14, 2026.
+
+Nội dung release:
+- Extension - Support display destination fee for transfer XCM (#4278)
+- Calculate exactly estimate fee for bridge step in liquid staking (#4803)
+- Disable all networks' switch to Manage Networks page (#4970)
+- Update for chain-list stable v0.2.127
+  - Support Transfer Alpha Token (#4900)
+  - Bridge native TAO <-> Subtensor EVM (#4901)
+  - Support bittensor on-chain swap (#4899)
+  - [Subtensor EVM] Add support for some alpha tokens
+
+
+## 1.3.77
+Build date: Apr 09, 2026.
+
+Features & Updates:
+- Improve Proxy account features (#4942)
+- Improve Multisig account feature  (#4963)
+- Turn off warning popup for transfers between PAH <> KAH (#4954)
+- Support stDOT LSD sunset (#4968)
+- Update for chainlist stable v0.2.126
+
+
+## 1.3.76
+Build date: Mar 20, 2026.
+
+Features & Updates:
+- Unable to turn network when no add correct API key (#4972)
+- Display token name and subnet ID for subnet tokens | Bittensor (#4892)
+- Support the newly updated Root staking feature | Bittensor (#4829)
+
+
+## 1.3.75
+Build date: Mar 17, 2026.
+
+Features & Updates:
+- Add user-configurable Subscan API Key in settings (#4965)
+
+
+## 1.3.74
+Build date: Feb 11, 2026.
+
+Features & Updates:
+- Support Multisig Account Phase 1 (#4855)
+
+
+## 1.3.73
+Build date: Jan 22, 2026.
+
+Features & Updates:
+- Update @subwallet-monorepos/subwallet-services-sdk 0.1.16 (#4957)
+- Remove the Crowdloans tab (#4920)
+
+Bug fixes:
+- Fix bug get Earning position parachain (#4950)
+
+
+## 1.3.72
+Build date: Jan 14, 2026.
+
+Features & Updates:
+- Integrate Proxy Account Support (#4725)
+- Update chain-list stable v0.2.123 (#4861)
+- Add support for USDC & stEWT (#639)
+- Migrate to ParaSpell V5 (#4908)
+
+
+## 1.3.71
+Build date: Dec 29, 2025.
+
+Features & Updates:
+- Improve token enabling (Round 2) (#4247)
+- Update libs for SubWallet Extensions (#4808)
+- Support Import from Trust Wallet to SubWallet (#4762)
+
+
+## 1.3.70
+Build date: Dec 11, 2025.
+
+Features & Updates:
+- Support OpenGov (Phase 1) (#4678)
+
+
 ## 1.3.69
 Build date: Dec 08, 2025.